extractor/pyproject.toml
[project]
name = "camper-extractor"
version = "0.1.0"
requires-python = ">=3.12"
description = "Feature extraction from Polish camper/van listings using OpenAI and Instructor"
dependencies = [
    "openai>=1.50.0",
    "instructor>=1.6.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "*.md",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88